Current status on refactoring code

+++ src/arm_core.c
@@ -6,7 +6,7 @@
 #include <stddef.h>
 #include <unistd.h>
 
-uint16_t arm_core_open(arm_core_state_t* state)
+uint32_t arm_core_open(arm_core_state_t* state)
 {
     if (state == NULL)
     {
@@ -31,7 +31,7 @@ uint16_t arm_core_open(arm_core_state_t* state)
     return 0;
 }
 
-uint16_t arm_core_close(arm_core_state_t* state)
+uint32_t arm_core_close(arm_core_state_t* state)
 {
     if (state == NULL)
     {
@@ -50,32 +50,32 @@ uint16_t arm_core_close(arm_core_state_t* state)
     return 0;
 }
 
-void arm_core_led_r_set(arm_core_state_t* state, uint16_t value)
+void arm_core_led_r_set(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
 }
 
-uint16_t arm_core_led_r_get(arm_core_state_t* state, uint16_t value)
+uint32_t arm_core_led_r_get(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
 }
 
-void arm_core_led_g_set(arm_core_state_t* state, uint16_t value)
+void arm_core_led_g_set(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
 }
 
-uint16_t arm_core_led_g_get(arm_core_state_t* state, uint16_t value)
+uint32_t arm_core_led_g_get(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
 }
 
-void arm_core_led_b_set(arm_core_state_t* state, uint16_t value)
+void arm_core_led_b_set(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET) = value;
 }
 
-uint16_t arm_core_led_b_get(arm_core_state_t* state, uint16_t value)
+uint32_t arm_core_led_b_get(arm_core_state_t* state, uint32_t value)
 {
-    *((uint16_t*)(((uint8_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
+    *((uint32_t*)(((uint32_t*)state->priv)) + NIT_ARM_CORE_LED_R_OFFSET);
 }
